Refactor parser into lexer and arithmetic modules
