+++ src/arith.h
@@ -1,0 +1,6 @@
+#ifndef ARITH_H
+#define ARITH_H
+
+long eval_arith(const char *expr);
+
+#endif /* ARITH_H */
